fix(auth): Split urlListener into web/react-native
Fixes #2808
